+++ 1.20.4/Common/src/main/java/fuzs/puzzleslib/api/core/v1/utility/NbtSerializable.java
@@ -2,6 +2,9 @@ package fuzs.puzzleslib.api.core.v1.utility;
 
 import net.minecraft.nbt.CompoundTag;
 
+import java.util.function.Function;
+import java.util.function.Supplier;
+
 /**
  * A basic template for an object that is serializable to Minecraft's nbt format.
  */
@@ -22,13 +25,28 @@ public interface NbtSerializable {
     void read(CompoundTag tag);
 
     /**
-     * Serialize the component to a {@link CompoundTag}.
+     * Encode the serializable instance to a {@link CompoundTag}.
      *
-     * @return the capability serialized to {@link CompoundTag}
+     * @return a tag containing the serialized instance
      */
     default CompoundTag toCompoundTag() {
         CompoundTag tag = new CompoundTag();
         this.write(tag);
         return tag;
+    }
+
+    /**
+     * Decode the serializable instance from a {@link CompoundTag}.
+     *
+     * @param factory supplier for a new empty instance
+     * @param <T>     serializable instance type
+     * @return a tag containing the serialized instance
+     */
+    static <T extends NbtSerializable> Function<CompoundTag, T> fromCompoundTag(Supplier<T> factory) {
+        return (CompoundTag tag) -> {
+            T nbtSerializable = factory.get();
+            nbtSerializable.read(tag);
+            return nbtSerializable;
+        };
     }
 }