+++ 1.20.4/Common/src/main/java/fuzs/puzzleslib/mixin/MixinConfigPluginImpl.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Set;
 
 public class MixinConfigPluginImpl implements IMixinConfigPlugin {
-    private static final Collection<String> DEVELOPMENT_MIXINS = Set.of("client.OptionsMixin");
+    private static final Collection<String> DEVELOPMENT_MIXINS = Set.of("server.DedicatedServerSettingsMixin", "server.EulaMixin");
 
     @Override
     public void onLoad(String mixinPackage) {