+++ 1.20.4/Fabric/src/main/java/fuzs/puzzleslib/fabric/impl/capability/data/FabricCapabilityKey.java
@@ -1,38 +1,52 @@
 package fuzs.puzzleslib.fabric.impl.capability.data;
 
-import dev.onyxstudios.cca.api.v3.component.ComponentAccess;
-import dev.onyxstudios.cca.api.v3.component.ComponentKey;
 import fuzs.puzzleslib.api.capability.v3.data.CapabilityComponent;
 import fuzs.puzzleslib.api.capability.v3.data.CapabilityKey;
 import fuzs.puzzleslib.impl.capability.GlobalCapabilityRegister;
+import net.fabricmc.fabric.api.attachment.v1.AttachmentTarget;
+import net.fabricmc.fabric.api.attachment.v1.AttachmentType;
 import net.minecraft.resources.ResourceLocation;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Objects;
 import java.util.function.Function;
+import java.util.function.Predicate;
+import java.util.function.Supplier;
 
+@SuppressWarnings("UnstableApiUsage")
 public abstract class FabricCapabilityKey<T, C extends CapabilityComponent<T>> implements CapabilityKey<T, C> {
-    private final ComponentKey<ComponentAdapter<T, C>> componentKey;
+    private final AttachmentType<C> attachmentType;
+    private final Predicate<Object> filter;
+    private final Function<T, Supplier<C>> factory;
 
-    public FabricCapabilityKey(ComponentKey<ComponentAdapter<T, C>> componentKey) {
-        this.componentKey = componentKey;
+    public FabricCapabilityKey(AttachmentType<C> attachmentType, Predicate<Object> filter, Supplier<C> factory) {
+        this.attachmentType = attachmentType;
+        this.filter = filter;
+        this.factory = (T holder) -> {
+            return () -> {
+                C capabilityComponent = factory.get();
+                Objects.requireNonNull(capabilityComponent, "capability component is null");
+                capabilityComponent.initialize((CapabilityKey<T, CapabilityComponent<T>>) this, holder);
+                return capabilityComponent;
+            };
+        };
         GlobalCapabilityRegister.register(this);
     }
 
     @Override
     public ResourceLocation identifier() {
-        return this.componentKey.getId();
+        return this.attachmentType.identifier();
     }
 
     @Override
     public C get(@NotNull T holder) {
         Objects.requireNonNull(holder, "holder is null");
-        if (holder instanceof ComponentAccess && this.isProvidedBy(holder)) {
-            ComponentAdapter<T, C> componentAdapter = this.componentKey.getNullable(holder);
-            Objects.requireNonNull(componentAdapter, "data is null");
-            C capabilityComponent = componentAdapter.getComponent();
+        if (holder instanceof AttachmentTarget attachmentTarget && this.isProvidedBy(holder)) {
+            C capabilityComponent = attachmentTarget.getAttachedOrCreate(this.attachmentType, this.factory.apply(holder));
             Objects.requireNonNull(capabilityComponent, "data is null");
+            // if the attachment is created from deserialization this has not been called yet
+            capabilityComponent.initialize((CapabilityKey<T, CapabilityComponent<T>>) this, holder);
             return capabilityComponent;
         } else {
             throw new IllegalArgumentException("Invalid capability holder: %s".formatted(holder));
@@ -41,11 +55,12 @@ public abstract class FabricCapabilityKey<T, C extends CapabilityComponent<T>> i
 
     @Override
     public boolean isProvidedBy(@Nullable Object holder) {
-        return holder instanceof ComponentAccess && this.componentKey.isProvidedBy(holder);
+        return holder instanceof AttachmentTarget && this.filter.test(holder);
     }
 
     @FunctionalInterface
-    public interface Factory<T, C extends CapabilityComponent<T>, K extends CapabilityKey<T, C>> extends Function<ComponentKey<ComponentAdapter<T, C>>, K> {
+    public interface Factory<T, C extends CapabilityComponent<T>, K extends CapabilityKey<T, C>> {
 
+        K apply(AttachmentType<C> attachmentType, Predicate<Object> filter, Supplier<C> factory);
     }
 }